oembed: ignore oembed results from mastodon
This adds a check in the oembed preview for cases where the oembed
is going to contain less information than the OpenGraph; in these
cases, the oembed is ignored and the OpenGraph data is retained.
At the moment, the only case in here is Mastodon posts, but it
could be expanded in the future.

Mastodon's oembed is JavaScript based and only contains stub HTML
that's intended to be filled in by client-side JavaScript that
synapse doesn't execute. It doesn't contain the actual post text.
The og:description, on the other hand, does contain the post text.
When synapse unilaterally prefers the oembed over OpenGraph, it
ends up accidentally filling in a generic and non-useful preview
in place of the actual post content.