oembed: ignore oembed results from Tumblr
Tumblr provides a generic oembed result that consists just of the
post URL. The og:description, on the other hand, contains an actual
abbreviated copy of the post, which is more useful.

This also adds a method to ignore other oembed results based on the
provider string. That wasn't possible for Mastodon, but it is for
Tumblr.